switched to float, fixed bug in proce table

+++ HotelBossEJB/ejbModule/cc/co/enricosartori/hotelboss/dto/Price.java
@@ -9,9 +9,9 @@ public class Price implements Serializable {
 	int price_id;
 	Date start_d;
 	Date end_d;
-	double fb;
-	double hb;
-	double bb;
+	float fb;
+	float hb;
+	float bb;
 	String state;
 	
 	public int getPrice_id() {
@@ -32,22 +32,22 @@ public class Price implements Serializable {
 	public void setEnd_d(Date endD) {
 		end_d = endD;
 	}
-	public double getFb() {
+	public float getFb() {
 		return fb;
 	}
-	public void setFb(double d) {
+	public void setFb(float d) {
 		this.fb = d;
 	}
-	public double getHb() {
+	public float getHb() {
 		return hb;
 	}
-	public void setHb(double d) {
+	public void setHb(float d) {
 		this.hb = d;
 	}
-	public double getBb() {
+	public float getBb() {
 		return bb;
 	}
-	public void setBb(double d) {
+	public void setBb(float d) {
 		this.bb = d;
 	}
 	public String getState() {
